requirements.txt
asgiref==3.4.1
Django==3.2.3
django-extensions==3.1.3
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Markdown==3.3.4
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.2
